refactor(banking): extract a shared useConnectFlow hook
The connect dialog and the inline connect flow shared ~190 lines of identical
state, institution fetching, authorize, and validation logic. Move it into a
useConnectFlow hook (plus the shared CONNECT_COUNTRIES); the two components
keep only their own chrome (dialog vs inline, haptics, back navigation).
